+++ src/api/potaApi.js
@@ -10,12 +10,12 @@ const axiosConfig = {
   baseURL: BASE_URL,
   timeout: 35000,
   headers: {
-    'User-Agent': 'RU-POTA-Bot/1.16.48 (Telegram Bot; Node.js)'
+    'User-Agent': 'RU-POTA-Bot/1.16.49 (Telegram Bot; Node.js)'
   }
 };
 
-if (process.env.TG_PROXY) {
-  const agent = new SocksProxyAgent(process.env.TG_PROXY, { keepAlive: true, keepAliveMsecs: 10000 });
+if (process.env.POTA_PROXY) {
+  const agent = new SocksProxyAgent(process.env.POTA_PROXY, { keepAlive: true, keepAliveMsecs: 10000 });
   axiosConfig.httpAgent = agent;
   axiosConfig.httpsAgent = agent;
 }